requirements.txt
FASTAPI>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
openpyxl>=3.1.2
dnspython>=2.4.2
httpx>=0.25.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
email-validator>=2.1.0
beautifulsoup4>=4.12.0
apscheduler>=3.10.4
lxml>=4.9.0
